pyproject.toml
[project]
name = "swc-shortener"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "environs>=15.0.1",
    "psycopg2>=2.9.12",
]


[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=6.0.3",
    "djangorestframework-stubs>=3.16.9",
    "factory-boy>=3.3.3",
    "mypy>=1.20.1",
    "ruff>=0.15.11",
]


[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "UP",  # pyupgrade
    "YTT", # flake8-2020
]
ignore = [
    "E203", # whitespace before ':' (conflicts with Black)
]
exclude = ["*/migrations/*.py", "*/*_sql.py"]

[tool.ruff.lint.isort]
sections = {"django" = ["django"]}
section-order = ["future", "standard-library", "third-party", "django", "first-party", "local-folder"]
lines-between-types = 0
lines-after-imports = 2

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main.main",
]

[tool.django-stubs]
django_settings_module = "src.swc_shortener.settings"
